requirements.txt
langchain
langchain-community

langchain-google-genai

google-api-python-client
requests
sentence-transformers
faiss-cpu
pypdf
